requirements.txt
google-cloud-aiplatform
google-auth
